64-bit clean and portable includes
- cgame syscalls carry pointer-sized arguments (intptr_t) and vmMain takes
  intptr_t arguments, matching the engine's VM_Call/VM_DllSyscall.
- FOFS/STOFS/LLOFS/CLOFS/NPCOFS use offsetof instead of casting a null
  pointer to int; the two remaining entity index <-> pointer casts in the
  savegame code go through intptr_t.  With pointer truncation promoted to an
  error the game and UI modules build as x86_64 without a single hit.
- The inline x86 "rep stosd" loops in g_objectives.cpp are plain C loops.
- #include paths use forward slashes and the exact on-disk filename case, so
  they resolve on case-sensitive filesystems.

+++ cgame/cg_main.cpp
@@ -1,10 +1,10 @@
 
 #include "cg_local.h"
-#include "fx_public.h"
-#include "..\client\vmachine.h"
+#include "FX_Public.h"
+#include "../client/vmachine.h"
 #include "cg_text.h"
-#include "..\game\characters.h"
-#include "..\game\speakers.h"
+#include "../game/characters.h"
+#include "../game/speakers.h"
 
 //NOTENOTE: Be sure to change the mirrored code in g_shared.h
 typedef	std::map< std::string, unsigned char, std::less<std::string>, std::allocator< unsigned char > >	namePrecache_m;
@@ -33,7 +33,7 @@ This is the only way control passes into the cgame module.
 This must be the very first function compiled into the .q3vm file
 ================
 */
-int vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7 ) {
+intptr_t vmMain( int command, intptr_t arg0, intptr_t arg1, intptr_t arg2, intptr_t arg3, intptr_t arg4, intptr_t arg5, intptr_t arg6, intptr_t arg7 ) {
 	switch ( command ) {
 	case CG_INIT:
 		CG_Init( arg0 );
